+++ src/App.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
-import ThemeSwitch from './components/ThemeSwitch.vue';
-import Nav from './components/Nav.vue';
+import NavBar from './components/NavBar.vue';
 import FileUpload from './components/FileUpload.vue';
 import FileItem from './components/FileItem.vue';
 import { useFileDataStore } from './stores/fileData';
@@ -9,6 +8,9 @@ import type { FileObj } from './types/file';
 import { computed, ref, watch } from 'vue';
 import { useImageCompression } from './composables/useImageCompression';
 import { useZipCompression } from './composables/useZipCompression';
+import { useI18n } from './hooks/useI18n';
+
+const { t } = useI18n();
 
 const fileDataStore = useFileDataStore();
 const { files } = storeToRefs(fileDataStore);
@@ -60,26 +62,25 @@ watch(anyUncompressed, async (newVal) => {
 
 <template>
   <header>
-    <Nav />
-    <ThemeSwitch />
-    <h1 class="title">{{ $t('translation.header') }}</h1>
+    <NavBar />
+    <h1 class="title">{{ t('translation.header') }}</h1>
   </header>
 
   <main>
     <div class="explanation-wrapper">
-      <p class="explanation">{{ $t('translation.subtitle') }}</p>
+      <p class="explanation">{{ t('translation.subtitle') }}</p>
       <a
         href="https://nomanssky.fandom.com/wiki/Special:Upload?multiple=true"
         role="button"
         target="_blank"
         rel="noopener noreferrer"
-        >{{ $t('translation.buttonwiki') }}</a
+        >{{ t('translation.buttonwiki') }}</a
       >
     </div>
-    <h2 class="subheading">{{ $t('translation.input') }}</h2>
+    <h2 class="subheading">{{ t('translation.input') }}</h2>
     <FileUpload />
 
-    <h2 class="subheading">{{ $t('translation.filelist') }}</h2>
+    <h2 class="subheading">{{ t('translation.filelist') }}</h2>
     <div class="buttons">
       <button
         :aria-busy="isCompressing"
@@ -87,7 +88,7 @@ watch(anyUncompressed, async (newVal) => {
         :disabled="!files.length || !anyUncompressed"
         @click="compressFiles"
       >
-        {{ files.length && !anyUncompressed ? $t('translation.allcompressed') : $t('translation.compress') }}
+        {{ files.length && !anyUncompressed ? t('translation.allcompressed') : t('translation.compress') }}
       </button>
       <a
         :aria-busy="isZipCompressing"
@@ -96,21 +97,21 @@ watch(anyUncompressed, async (newVal) => {
         role="button"
         download
       >
-        {{ $t('translation.downloadzip') }}
+        {{ t('translation.downloadzip') }}
       </a>
       <button
         :disabled="!files.length"
         class="secondary"
         @click="files = []"
       >
-        {{ $t('translation.clearlist') }}
+        {{ t('translation.clearlist') }}
       </button>
     </div>
     <div class="file-list">
       <FileItem
         v-for="file in files"
-        :key="file.id"
         :file-obj="file"
+        :key="file.id"
         @remove="removeItem(file)"
       />
     </div>
