+++ src/components/ThemeSwitch.vue
@@ -1,11 +1,15 @@
 <script setup lang="ts">
+import { useI18n } from '../hooks/useI18n';
+
+const { t } = useI18n();
+
 // determine preferred theme and update the html element with the respective tag
 const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
 switchTheme(prefersDark ? 'dark' : 'light');
 
 function switchTheme(theme: string | undefined = undefined) {
   const currentTheme = document.documentElement.dataset.theme;
-  const computedNewTheme = currentTheme == 'dark' ? 'light' : 'dark';
+  const computedNewTheme = currentTheme === 'dark' ? 'light' : 'dark';
   const newTheme = theme ?? computedNewTheme;
   document.documentElement.dataset.theme = newTheme;
 }
@@ -14,13 +18,12 @@ function switchTheme(theme: string | undefined = undefined) {
 <template>
   <div style="text-align: right">
     <button
-      role="button"
       class="themeswitcher"
       id="themeSwitch"
-      @click="switchTheme()"
       style="width: auto"
+      @click="switchTheme()"
     >
-      {{ $t('translation.switchtheme') }}
+      {{ t('translation.switchtheme') }}
     </button>
   </div>
 </template>
